eth mix digest
