+++ go.mod
@@ -13,6 +13,7 @@ require (
 	github.com/grpc-ecosystem/grpc-gateway v1.16.0
 	github.com/irisnet/irismod v1.4.1-0.20210823060730-52fb3e67d950
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
+	github.com/ontio/ontology v1.11.1-0.20200812075204-26cf1fa5dd47
 	github.com/petermattis/goid v0.0.0-20180202154549-b0b1615b78e5 // indirect
 	github.com/pkg/errors v0.9.1
 	github.com/polynetwork/poly v1.7.3-0.20210817091119-a76693e9abd0
